refactor(ui): split NetworkDiscoveryCard into types, helpers, scan hook
Drop NetworkDiscoveryCard.tsx from 1375 to 398 lines, well under the
800-line red-flag budget. Three sibling files share the load.

- networkDiscoveryCardTypes.ts (219): every discovery + SNMP +
  port-scan + deep-scan type, plus DiscoveredDevice / DiscoveryStatus /
  NetworkDiscoveryData themselves. Public type names re-exported from
  NetworkDiscoveryCard.tsx so existing consumers still import from
  './NetworkDiscoveryCard'
- NetworkDiscoveryCardHelpers.tsx (393): formatLastSeen,
  calculateNetworkAddress, subnetList, categorizeDevices,
  discoverySummary, and COMMON_PORTS
- useNetworkDiscoveryAutoScan.ts (394): port-scan + vulnerability-scan
  automation - settings fetch, manual deep-scan callback, and the two
  auto-scan effects with their tracked-set refs to prevent duplicates

Pure relocation - the lowercase <discoverySummary> / <subnetList> JSX
tags are kept as they were in the original file. No behavior change.
